+++ core/src/test/java/com/asml/apa/wta/core/ConfigReaderTest.java
@@ -8,7 +8,7 @@ import com.asml.apa.wta.core.utils.WtaUtils;
 import java.util.HashMap;
 import org.junit.jupiter.api.Test;
 
-public class ConfigReaderTest {
+class ConfigReaderTest {
   @Test
   void readsConfigFileCorrectly() {
     RuntimeConfig cr = WtaUtils.readConfig("src/test/resources/testConfig.json");
@@ -21,6 +21,9 @@ public class ConfigReaderTest {
         put("f2", "v2");
       }
     });
+    assertThat(cr.getLogLevel()).isEqualTo("DEBUG");
+    assertThat(cr.isDoConsoleLog()).isFalse();
+    assertThat(cr.isDoFileLog()).isTrue();
   }
 
   @Test
@@ -35,6 +38,9 @@ public class ConfigReaderTest {
         put("f2", "v2");
       }
     });
+    assertThat(cr.getLogLevel()).isEqualTo("DEBUG");
+    assertThat(cr.isDoConsoleLog()).isFalse();
+    assertThat(cr.isDoFileLog()).isTrue();
   }
 
   @Test
@@ -52,5 +58,22 @@ public class ConfigReaderTest {
           WtaUtils.readConfig("src/test/resources/testConfigInvalid.json");
         })
         .isInstanceOf(IllegalArgumentException.class);
+  }
+
+  @Test
+  void readsConfigFileWhereLogSettingIsNotThere() {
+    RuntimeConfig cr = WtaUtils.readConfig("src/test/resources/testConfigNoLogSettings.json");
+    assertThat(cr.getAuthor()).isEqualTo("Test Name");
+    assertThat(cr.getDomain()).isEqualTo("Test Domain");
+    assertThat(cr.getDescription()).isEqualTo("Test Description");
+    assertThat(cr.getEvents()).isEqualTo(new HashMap<String, String>() {
+      {
+        put("f1", "v1");
+        put("f2", "v2");
+      }
+    });
+    assertThat(cr.getLogLevel()).isEqualTo("INFO");
+    assertThat(cr.isDoConsoleLog()).isTrue();
+    assertThat(cr.isDoFileLog()).isTrue();
   }
 }